Move and improve memory profiling to benchmarks folder
