+++ benchmarks/memory/use_case/failure/with_transitions/u_case_v2-6-0.rb
@@ -7,31 +7,24 @@ gemfile do
 
   gem 'memory_profiler'
 
-  gem 'u-case', '~> 3.1.0'
+  gem 'u-case', '~> 2.6.0'
 end
 
-Micro::Case.config do |config|
-  config.enable_transitions = true
-end
+# Micro::Case::Result.disable_transition_tracking
 
 class Multiply < Micro::Case
   attributes :a, :b
 
   def call!
     if a.is_a?(Numeric) && b.is_a?(Numeric)
-      Success(result: { number: a * b })
+      Success { { number: a * b } }
     else
       Failure(:invalid_data)
     end
   end
 end
 
-SYMBOL_KEYS = { a: nil, b: 2 }
-STRING_KEYS = { 'a' => 1, 'b' => '' }
+Multiply.call(a: nil, 'b' => 2)
 
-report = MemoryProfiler.report do
-  Multiply.call(SYMBOL_KEYS)
-  Multiply.call(STRING_KEYS)
-end
-
+report = MemoryProfiler.report { Multiply.call(a: nil, 'b' => 2) }
 report.pretty_print
