+++ benchmarks/memory/use_case/failure/without_transitions/u_case_v2-0-0.rb
@@ -1,0 +1,28 @@
+require 'bundler/inline'
+
+gemfile do
+  source 'https://rubygems.org'
+
+  ruby '>= 2.4.0'
+
+  gem 'memory_profiler'
+
+  gem 'u-case', '~> 2.0.0'
+end
+
+class Multiply < Micro::Case
+  attributes :a, :b
+
+  def call!
+    if a.is_a?(Numeric) && b.is_a?(Numeric)
+      Success { { number: a * b } }
+    else
+      Failure(:invalid_data)
+    end
+  end
+end
+
+Multiply.call(a: nil, 'b' => 2)
+
+report = MemoryProfiler.report { Multiply.call(a: nil, 'b' => 2) }
+report.pretty_print
