+++ benchmarks/memory/use_case/success/with_transitions/u_case_v3-1-0.rb
@@ -7,7 +7,11 @@ gemfile do
 
   gem 'memory_profiler'
 
-  gem 'u-case', '~> 2.0.0'
+  gem 'u-case', '~> 3.1.0'
+end
+
+Micro::Case.config do |config|
+  config.enable_transitions = true
 end
 
 class Multiply < Micro::Case
@@ -15,19 +19,17 @@ class Multiply < Micro::Case
 
   def call!
     if a.is_a?(Numeric) && b.is_a?(Numeric)
-      Success { { number: a * b } }
+      Success(result: { number: a * b })
     else
       Failure(:invalid_data)
     end
   end
 end
 
-SYMBOL_KEYS = { a: nil, b: 2 }
-STRING_KEYS = { 'a' => 1, 'b' => '' }
+Multiply.call(a: 2, 'b' => 2)
 
 report = MemoryProfiler.report do
-  Multiply.call(SYMBOL_KEYS)
-  Multiply.call(STRING_KEYS)
+  Multiply.call(a: 2, 'b' => 2)
 end
 
 report.pretty_print
