+++ benchmarks/memory/use_case/success/without_transitions/u_case_v3-1-0.rb
@@ -1,0 +1,35 @@
+require 'bundler/inline'
+
+gemfile do
+  source 'https://rubygems.org'
+
+  ruby '>= 2.4.0'
+
+  gem 'memory_profiler'
+
+  gem 'u-case', '~> 3.1.0'
+end
+
+Micro::Case.config do |config|
+  config.enable_transitions = false
+end
+
+class Multiply < Micro::Case
+  attributes :a, :b
+
+  def call!
+    if a.is_a?(Numeric) && b.is_a?(Numeric)
+      Success(result: { number: a * b })
+    else
+      Failure(:invalid_data)
+    end
+  end
+end
+
+Multiply.call(a: 2, 'b' => 2)
+
+report = MemoryProfiler.report do
+  Multiply.call(a: 2, 'b' => 2)
+end
+
+report.pretty_print
